Implement Hex Decode
Hexadecimal (base16) is a convenient way to represent binary data, this is because 4 bits can be represented by one character.
As such one byte (8 bits) can be represented with two characters, i.e. 11111111 == FF.

The bitwise left shift operator `<<` shifts the bits to the left, padding the new bytes with zero e.g. 2 << 4 == 0010000.
The or operator `|` is a logical `OR` (of course).

To get the raw bytes we can just parse the hex data in pairs, parsing each character to an integer. We can apply a left shift of 4 to the first value and or it with second value to get the byte represented by the pair. We get a nice reward "I'm killing your brain like a poisonous mushroom"