Implement Base64 Encode
Base 64 can represent 3 bytes in 4 characters. This is because each character can have 64 possible values, which is the same as 6 bits (2^6). As a byte is 8 bits, we must find the lowest common multiple of 8 and 6 (24). So we must work with chunks of 3 bytes and produce 4 characters.

W can use a bit mask (logical AND) combined with a right shift to get the N most significant bits. The mask to get the N most significant bits will be (2^N) - 1. To keep the 6 most significant bits we would use (2^6) - 1 = 63 = 0x3f.

As such we parse our 3 bytes to a single 32 bit int by shifting left and ORing them (same as hex decoding). Then we can use the mask we calculated above combined with right shifts to convert the 32bit value into 4 6bit values.  Use the 4 6bit values to lookup the character in our conversion table. The remaining bytes are handled with ugly branches, but using the same operators.