Set 1 Challenge 6: Implement Hamming distance
Counting Set Bits

This is Brian Kernighan’s counting method. It works because the least significant bit is set if number is odd. Therefore to find how many bits are set in a byte B we can repeatedly clear the least significant bit by ANDing B & B - 1 and count how many iterations it takes to reach 0.